% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCA.R
\name{BCA}
\alias{BCA}
\title{BCA Nonparametric confidence intervals}
\usage{
BCA(df, pval, B)
}
\arguments{
\item{df}{a data.frame}
\item{pval}{significance level defined by the user}
\item{B}{an integer value, specify the number of resamples}
}
\value{
It returns a customized table using the expss package. The table
`b_BCA` for each model parameters (row): original (opt), low (lw)
and high (up) CI, bias and coefficient of variation (CV): Opt, IClw,
ICup, BClw, BCup, BCAlw, BCAup, Bias, CV.
}
\description{
This function takes a data frame (df), requires the significance
value (pval) and calculates the nonparametric confidence
intervals (CI): Symmetric (IC), Bias-corrected (BC) and BC accelerated (BCA)

}
\examples{
  \dontrun{
    df <- tmp
    pval <- 0.05
    B <- 2000
    ejemplo <- BCA(df, pval, B)
  }
}
