
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_Barplot.R
\name{fn_Barplot}
\alias{fn_Barplot}
\title{fn_Barplot}
\usage{
fn_Barplot(df, i, densidad, angulo)
}
\arguments{
\item{df}{A list of matrices with frequencies values.}

\item{i}{A numeric value of the ith object in the list.}

\item{densidad}{a vector that specifies the shading lines density (in lines
per inch). By default, it is NULL, which means no shading lines. For
more details, see the help of the barplot() function}

\item{angulo}{a vector that specifies the slope of shading lines. For
more details, see the help of the barplot() function}
}
\value{
It returns a simple barplot that can be upgraded according to needs.
}
\description{
A very simple barplot is generated with no labels or marks for the X-axis,
    which serves as the basis for those that will be generated through the
    analysis, whether they are simple or compound figures
}
\examples{
  \dontrun{
For example, for two sampled years i = 2 and six factor levels in the
 analyzed variable.

      df <- tmp
      i <- 2
      densidad <- c(10,10,30,30,30,10)
      angulo <- c(3,30,60,90,120,150)

      fn_Barplot(df,i,density,angle)
    }

}
