% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_Boxplot.R
\name{fn_Boxplot}
\alias{fn_Boxplot}
\title{fn_Boxplot}
\usage{
fn_Boxplot(val, fac, df, yrange)
}
\arguments{
\item{val}{A column containing the numeric values.}

\item{fac}{The factor column (grouping column).}

\item{df}{The data frame containing the data.}

\item{yrange}{A vector with the minimum and maximum values of the y-axis.}
}
\value{
A multiple boxplots which can be upgraded according to needs.
}
\description{
A very simple multiple boxplots (more than one level of factor) is generated
with no labels or marks for the X-axis. It serves as the basis for the figures that will be generated through the analysis, whether they are simple or complex.
}
\examples{
  \dontrun{
For example, in the tmp dataframe, the variable to analyze is in the LT
     column and the factors in the Phases column.

      val <- tmp$LT
      yrange <- c(20,300)
      fac <- as.factor(tmp$Phase)

      fn_Boxplot(val,fac,df,ylim)
    }
}
