% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_coincide.R
\name{fn_coincide}
\alias{fn_coincide}
\title{fn_coincide}
\usage{
fn_coincide(N_Phase, df, Icut, Coincide)
}
\arguments{
\item{N_Phase}{A numerical value indicating the number of factor levels}

\item{df}{A data frame with the values of each individual (row) by variables
(columns)}

\item{Icut}{A numerical value is the threshold value to separate into two
groups}

\item{Coincide}{A table filled with NA, with N+1 rows, one for each factor
level plus 1 for the total, and four columns, the first with the name of
factor levels and the rest for the categories: C, MI and IM}
}
\value{
Return a filled table with the number of coincidences and discrepancies by factor
levels and the total (rows) by category (columns): Coincidence (C), mature by
immature (MI) and immature by mature (IM).
}
\description{
The coincidence between microscopy classification and GSI is analyzed. The count
is based on the number of pairs that coincide (C) and the number of pairs that
are misqualified: mature by immature (MI) and immature by mature (IM).
}
\examples{
  \dontrun{
  N_Phase <- 6
  df <- tmp
  Icut <- 2
  Coincide <- data.frame(matrix(NA, ncol = 4, nrow = as.integer(N_Phase+1)))
  ejemplo = fn_coincide(N_Phase, df, Icut, Coincide)

    }
}
