% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.R
\name{fn_freq}
\alias{fn_freq}
\title{fn_freq}

\usage{
fn_freq(df, Imin, Imax, bin)
}
\arguments{
\item{df}{A data frame}

\item{Imin}{Minimum value to be defined for the x-axis}

\item{Imax}{Maximum value to be defined for the x-axis}

\item{bin}{A bin value to define the width of the class interval}
}
\value{
Table with the absolute, cumulative and relative cumulative
 frequency by interval and class mark.
}
\description{
The interval and class mark are defined and the absolute, cumulative and
 relative cumulative frequency is calculated in a table.
}
\examples{
  \dontrun{
    df <- tmp
    Imin <- 1
    Imax <- 30
    bin <- 1
    ejemplo <-  fn_freq(df, Imin, Imax, bin)
  }
}
