% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_hsd.R
\name{fn_hsd}
\alias{fn_hsd}
\title{fn_hsd}
\usage{
fn_hsd(hsd_out, Phases_legend)
}
\arguments{
\item{hsd_out}{A TukeyHSD test table, comes from the result of the
Tukey_HSD function}
\item{Phases_legend}{A character vector that has the name or number of maturity stages defined by the user}
}
\value{
A vector with the names of the phases arranged in descending order of the average response.
}
\description{
The factor column resulting from the TukeyHSD test is sorted by its row name,
these are renamed with the name of the phases and finally, they are sorted,
in decreasing way, by the average response values (column 1).
}
\examples{
  \dontrun{
    hsd_out <- tmp
    Phases_legend <- c("I", "II", "III", "IV", "V", "VI")

    ejemplo <- fn_hsd(hsd_out,Phases_legend)
  }
}
