% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_Idioma.R
\name{fn_Idioma}
\alias{fn_Idioma}
\title{fn_Idioma}
\usage{
fn_Idioma(idioma, language)
}
\arguments{
\item{idioma}{An integer value that defines the language of the labels.}

\item{language}{A list containing the labels of each language.}
}
\value{
**etiquetas** A Vector is defined by its labels, which include:
    total length, frequency, year, month, Gonadosomatic Index, maturity phase,
    and residuals. If more tags are needed, it must be redefined.
    They should be changed here in the function.
}
\description{
The figure labels are defined as either English (1) or Spanish (2).
}
\examples{
  \dontrun{
  Labels are in English
    idioma <-   1
        fn_Idioma(idioma,language)
  Labels are in Spanish
    idioma <-   2
        fn_Idioma(idioma,language)
  }
}
