% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IC_plot.R
\name{IC_plot}
\alias{IC_plot}
\title{IC_plot}
\usage{
IC_plot(x, y, yc, MClw, MCup, CIl, CIh, xl, yl)
}
\arguments{
\item{x}{A vector of the X-axis values}

\item{y}{A vector of the y-axis values}

\item{yc}{A vector with Y values calculated from the original data}

\item{MClw}{Lower value to be defined for the x-axis}

\item{MCup}{Upper value to be defined for the x-axis}

\item{CIl}{A table with the lower value of nonparametric confidence interval}

\item{CIh}{A table with the upper value of nonparametric confidence interval}

\item{xl}{A label for the X-axis}

\item{yl}{A label for the y-axis}
}
\value{
It returs a graphic representation of the cumulative percentage of mature females
as a function of length (points). The fitted model (black line) and its 95 percent
confidence interval (gray shaded area) are included; the horizontal dashed line
represents the cumulative ratio of 0.5; The vertical dashed line indicates the L50
value and the outer lines the 95 percent of confidence interval.

}
\description{
Cumulative percentage of mature females against length. Fitting model for the
   estimation of the L50 values and its confidence interval.
}
\examples{
  \dontrun{
 IC_plot(x,y,yc,MClw,MCup,CIl,CIh,xl,yl)
  }
}
