% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Label_Phases.R
\name{Label_Phases}
\alias{Label_Phases}
\title{Label_Phases}
\usage{
Label_Phases(genero, idioma, Phases)
}
\arguments{
\item{genero}{An integer value to define the labels gender}
\item{idioma}{An integer value to define the labels language}
\item{Phases}{An integer value to define the labels of the phases of maturity}
}
\value{
Phases_legend
}
\description{
The function returns the maturity phase label that will be used in tables or
figures. It returns a 'Phases_legend' vector in roman numeral or name in English
or Spanish according to the user's choice.
}
\details{
Arguments are 'genero' (Gender),'idioma' (Lenguage) and Phases (maturity phases)
    genero: Value *1* selects Females and *2* Males.
    idioma: Value *1* selects English language and *2* Spanish language.
    Phases: Value *1* selects the roman numeral and *2* phases name.
}
\examples{
  \dontrun{
    Label_Phases(1, 1, 1) # returns roman numeral for females
    Label_Phases(1, 2, 1) # returns roman numeral for females
    Label_Phases(1, 1, 2) # returns phases name in English for females
    Label_Phases(1, 2, 2) # returns phases name in Spanish for females
    Label_Phases(2, 1, 1) # returns roman numeral for males
    Label_Phases(2, 2, 1) # returns roman numeral for males
    Label_Phases(2, 1, 2) # returns phases name in English for males
    Label_Phases(2, 2, 2) # returns phases name in Spanish for males
  }
}
