% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelo.R
\name{modelo}
\alias{modelo}
\title{modelo}
\usage{
modelo(x,y,mod,bb,bmin,bmax)
}
\arguments{
\item{x}{independent variable.}
\item{y}{dependent variable.}
\item{mod}{A nonlinear model formula including variables and parameters. Will
be coerced to a formula if necessary.}
\item{bb}{A list or numeric vector of starting estimates. If par is a
list, then each element must be of length 1.}
\item{bmin}{A numeric vector of lower bounds on each parameter. If not
given, the default lower bound for each parameter is set to -Inf.}
\item{bmax}{A numeric vector of upper bounds on each parameter. If not
given, the default upper bound for each parameter is set to Inf.}
}

\value{
Model settings list
}
\description{
Fit the model using the nlsLM function. For more details see the function
documentation: nlsLM() from minpack.lm (version 1.2-4).
}
\examples{
  \dontrun{
 modelo(x,y,mod,bb,bmin,bmax)
  }
}
