% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strb.R
\name{strb}
\alias{strb}
\title{strb}
\usage{
strb(b3,b4,b5,b6,b7,x,y,mod,modelo, bl, bu, xl, yl)
}
\arguments{
\item{b3}{The initial value for the medium asymptote. Valid only for double
sigmoid model, NA is assigned for the simple model}

\item{b4}{The initial value for the coordinate at the inflection point for
simple sigmoid model. For the double sigmoid model, it corresponds
to the first inflection point}

\item{b5}{The initial value for the amplitudes of the sigmoid, for the double
sigmoid model corresponds to the amplitude of the first sigmoid.
b5= L95a-b4 ,L95a is the lengths at 95 percent maturity for the
complete sigmoid (single model) or the first sigmoid (double model)}

\item{b6}{The coordinate at the second inflection point. Valid only for double
sigmoid model, NA is assigned for the simple model}

\item{b7}{The initial value for the amplitude of the second sigmoid. Valid
only for double sigmoid model, NA is assigned for the simple model.
b7= L95b-b6 ,L95b is the lengths at 95 percent maturity from the
second distribution}

\item{x}{A vector of the X-axis values}

\item{y}{A vector of the Y-axis values}

\item{mod}{A nonlinear model formula including variables and parameters}

\item{modelo}{function from Repbio, to fit the specified model (mod)}

\item{bl}{Lower (bl) weight values apply to each parameter (bi)}

\item{bu}{Upper (bu) weight values apply to each parameter (bi).}

\item{xl}{Title label for X axis}

\item{yl}{Title label for Y axis}
}
\value{
1) A plot with the observed and fitted values

2) Model settings list
}
\description{
This function helps in the visual search of the starting input parameters
for nlsLM, to fit the specified model (mod), using a graph for visual
adjustment of model parameters (Trial and Error Strategy).
}
\examples{
  \dontrun{
 strb(b3,b4,b5,b6,b7,x,y,mod,modelo)
  }
}
