% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tukey_HSD.R
\name{Tukey_HSD}
\alias{Tukey_HSD}
\title{Tukey_HSD}
\usage{
Tukey_HSD(Prob, pval, HSD_main, HSD_trt,
          fn_hsd, etiquetas, formula,
          criterio, Phases_legend)
}
\arguments{
\item{Prob}{A summary from ANOVA-1 function.}
\item{pval}{The significance level (user defined).}
\item{HSD_main}{A mine title (user defined).}
\item{HSD_trt}{Constant( only y=model) or vector treatment applied to each
experimental unit.}
\item{fn_hsd}{This function is included in this package Repbio, it
returns the phase names in descending order by the average of the
response values.}
\item{etiquetas}{A data frame containing the user-defined labels that will
be used throughout the routine}
\item{formula}{A formula notation y~f where y is the response (independent)
variable and f is a factor (categorical) variable representing group
membership.}
\item{criterio}{Criteria for separating active reproductive organisms.}
\item{Phases_legend}{A character vector with the name or number of maturity
stages, as defined by the user.}
}
\value{
1) A TukeyHSD test table hsd_out saved as txt file as: Tukey
   plus the name of the model variables, acronym criteria established to
   classify active females to reproduction (criterio) and the txt extension.

2) A TukeyHSD contrast plot saved as pdf file as: Tukey
   plus the name of the model variables, acronym criteria established to
   classify active females to reproduction (criterio) and the txt extension
   Note: criterio is defined outside of this function.
}
\description{
This function performs multiple comparisons of treatments by means of Tukey
only if the p value of the ANOVA_1 test is equal to or less than pval
defined by the user. HSD.test() came from the agricolae package.
}
\details{
the TukeyHSD test (Tukey Honest Significant Differences) is performed.
}
\examples{
  \dontrun{
   pval <- 0.05
   HSD_main <- LT/nPhase
   HSD_trt <- Phase
   Tukey_HSD(Prob, pval, HSD_main, HSD_trt, fn_hsd, etiquetas,
   formula, criterio, Phases_legend)
  }
}

\seealso{
fn_hsd() from agricolae (version 1.3-7).
}
